Fixes for Linux build and jpeg loading

+++ lib/engine_graphics/include/bitmap/vsx_bitmap_transform.h
@@ -125,7 +125,7 @@ public:
       GLfloat* data2 = (GLfloat*)malloc(sizeof(GLfloat) * width * height * channels);
       int dy = 0;
       size_t sxbpp = width * channels;
-      for (size_t y = height-1; y >= 0; --y)
+      for (int y = height-1; y >= 0; --y)
       {
         for (unsigned long x = 0; x < width * channels; ++x)
           data2[dy*sxbpp + x] = ((GLfloat*)bitmap->data_get())[y*sxbpp + x];
@@ -139,7 +139,7 @@ public:
     unsigned char* data2 = (unsigned char*)malloc(width * height * channels);
     size_t dy = 0;
     size_t stride_x = width * channels;
-    for (size_t y = height-1; y >= 0; --y)
+    for (int y = height-1; y >= 0; --y)
     {
       size_t dysxbpp = dy * stride_x;
       size_t ysxbpp = y * stride_x;
